requirements.txt
# === Core ===
streamlit>=1.36.0
python-dotenv>=1.0.1
openai>=1.40.0

# === PDF / Parsing ===
PyMuPDF>=1.24.5
pandas>=2.2.2
numpy>=1.26.4
pillow>=10.3.0

# === Viz ===
matplotlib>=3.8.4

# === Search (RAG 등에서 사용 시) ===
rapidfuzz>=3.9.3

# === OCR 폴백(이미지/스캔 표 대응) ===
pytesseract>=0.3.10
opencv-python-headless>=4.9.0.80
